refactored csep1 tests into separate module and updated w-test to work with other csep classes.
